+++ web-console/src/lib/functions/common/array.ts
@@ -13,8 +13,8 @@ export const partition = <T>(arr: T[], predicate: (v: T, i: number, ar: T[]) => 
     [[], []] as [T[], T[]]
   )
 
-export function inUnion<T extends readonly string[]>(union: T, val: string): val is T[number] {
-  return union.includes(val)
+export function inUnion<T extends readonly string[]>(union: T, val: string | unknown): val is T[number] {
+  return typeof val === 'string' && union.includes(val)
 }
 
 export function invariantUnion<T extends readonly string[]>(union: T, val: string): asserts val is T[number] {