+++ web-console/src/lib/types/connectors.ts
@@ -1,3 +1,5 @@
+import { AttachedConnector, ConnectorDescr, ProgramDescr } from '$lib/services/manager'
+
 export enum ConnectorType {
   KAFKA_IN = 'KafkaIn',
   KAFKA_OUT = 'KafkaOut',
@@ -11,4 +13,14 @@ export enum Direction {
   INPUT = 'input',
   OUTPUT = 'output',
   INPUT_OUTPUT = 'input_output'
+}
+
+export type IONodeData = {
+  connector: ConnectorDescr
+  ac: AttachedConnector
+}
+
+export type ProgramNodeData = {
+  label: string
+  program: ProgramDescr
 }